+++ executor/builder.go
@@ -41,7 +41,6 @@ import (
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/ranger"
 	"github.com/pingcap/tipb/go-tipb"
-	log "github.com/sirupsen/logrus"
 	"golang.org/x/net/context"
 )
 
@@ -206,7 +205,6 @@ func (b *executorBuilder) buildShowDDLJobs(v *plan.ShowDDLJobs) Executor {
 }
 
 func (b *executorBuilder) buildCheckIndex(v *plan.CheckIndex) Executor {
-	log.Warnf("executor ---------------------------------")
 	readerExec, err := buildNoRangeIndexLookUpReader(b, v.IndexLookUpReader)
 	if err != nil {
 		b.err = errors.Trace(err)
@@ -217,7 +215,9 @@ func (b *executorBuilder) buildCheckIndex(v *plan.CheckIndex) Executor {
 
 	e := &CheckIndexExec{
 		baseExecutor: newBaseExecutor(b.ctx, v.Schema(), v.ExplainID()),
-		tables:       v.Tables,
+		dbName:       v.DBName,
+		tableName:    readerExec.table.Meta().Name.L,
+		idxName:      v.IdxName,
 		is:           b.is,
 		src:          readerExec,
 	}
