+++ plan/optimizer.go
@@ -24,7 +24,6 @@ import (
 	"github.com/pingcap/tidb/privilege"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/terror"
-	// log "github.com/sirupsen/logrus"
 )
 
 // AllowCartesianProduct means whether tidb allows cartesian join without equal conditions.
@@ -82,14 +81,6 @@ func Optimize(ctx sessionctx.Context, node ast.Node, is infoschema.InfoSchema) (
 	if logic, ok := p.(LogicalPlan); ok {
 		return doOptimize(builder.optFlag, logic)
 	}
-	//	if checkTable, ok := p.(*CheckTable); ok {
-	//		// TODO: make CheckTable has LogicalPlan
-	//		err := builder.buildCheckTable(checkTable)
-	//		if err != nil {
-	//			return nil, errors.Trace(err)
-	//		}
-	//		p = checkTable
-	//	}
 	if execPlan, ok := p.(*Execute); ok {
 		err := execPlan.optimizePreparedPlan(ctx, is)
 		return p, errors.Trace(err)
